Se agrego el sistema de validacion y se hizo cambios en el diseño de todas las paginas.
Se implemento fortify como sistema de autenticacion.

Se modifico el archivo de ruta web para agregar una funcion de grupo que aloja a distintas rutas

+++ app/Http/Controllers/Secretaria/DoctorSecretariaController.php
@@ -7,6 +7,7 @@ use App\Models\Consultorio;
 use App\Models\Doctor;
 use App\Models\Vistas\Vw_doctor_consultorio;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class DoctorSecretariaController extends Controller
@@ -18,13 +19,14 @@ class DoctorSecretariaController extends Controller
      */
     public function index()
     {
+        if(Gate::authorize('loginSecretaria')){
 
-        $doctor=Vw_doctor_consultorio::all();
-        $vista=view('vista_paginas.secretaria.doctor.vista_lista_doctor_secretaria')
-        ->with('listado_doctores', $doctor);
+            $doctor=Vw_doctor_consultorio::all();
+            $vista=view('vista_paginas.secretaria.doctor.vista_lista_doctor_secretaria')
+                ->with('listado_doctores', $doctor);
 
-        return $vista;
-
+            return $vista;
+        }
     }
 
     /**
@@ -34,12 +36,15 @@ class DoctorSecretariaController extends Controller
      */
     public function create()
     {
-        $consultorio=Consultorio::all();
-        $vista=view('vista_paginas.secretaria.doctor.crear_doctor_secretaria')
-        ->with('consultorios', $consultorio);
 
-        return $vista;
+       if(Gate::authorize('loginSecretaria')){
 
+           $consultorio=Consultorio::all();
+           $vista=view('vista_paginas.secretaria.doctor.crear_doctor_secretaria')
+               ->with('consultorios', $consultorio);
+
+           return $vista;
+       }
     }
 
     /**
@@ -51,38 +56,41 @@ class DoctorSecretariaController extends Controller
     public function store(Request $request)
     {
 
-        $request->validate([
-            'inputImgPerfil' => ['image'],
-            'consultorio' => 'required',
-            'nombre' => 'required',
-            'apellido' => 'required',
-            'email' => ['required', 'email'],
-            'telefono' => ['required', 'integer'],
-            'celular' => ['required', 'integer'],
-            'sexo' => 'required',
-        ]);
+        if(Gate::authorize('loginSecretaria')){
 
-        $doctor=new Doctor();
-        $doctor->id_consultorio=$request->input('consultorio');
-        $doctor->nombre=$request->input('nombre');
-        $doctor->apellidos=$request->input('apellido');
-        $doctor->usuario=$request->input('nombre').rand();
-        $doctor->email=$request->input('email');
-        $doctor->telefono=$request->input('telefono');
-        $doctor->celular=$request->input('celular');
-        $doctor->sexo=$request->input('sexo');
-        $doctor->horarios=$request->input('horarios');
-        $doctor->rol="Doctor";
+            $request->validate([
+                'inputImgPerfil' => ['image'],
+                'consultorio' => 'required',
+                'nombre' => 'required',
+                'apellido' => 'required',
+                'email' => ['required', 'email'],
+                'telefono' => ['required', 'integer'],
+                'celular' => ['required', 'integer'],
+                'sexo' => 'required',
+            ]);
 
-        if($request->hasFile('inputImgPerfil')){
-            $doctor->profile_photo_path=$request->file('inputImgPerfil')->store('public/fotos_perfil');
+            $doctor=new Doctor();
+            $doctor->id_consultorio=$request->input('consultorio');
+            $doctor->nombre=$request->input('nombre');
+            $doctor->apellidos=$request->input('apellido');
+            $doctor->usuario=$request->input('nombre').rand();
+            $doctor->email=$request->input('email');
+            $doctor->telefono=$request->input('telefono');
+            $doctor->celular=$request->input('celular');
+            $doctor->sexo=$request->input('sexo');
+            $doctor->horarios=$request->input('horarios');
+            $doctor->rol="Doctor";
+
+            if($request->hasFile('inputImgPerfil')){
+                $doctor->profile_photo_path=$request->file('inputImgPerfil')->store('public/fotos_perfil');
+            }
+
+            $doctor->save();
+
+
+            return redirect()->route('secretaria.doctor.edit',$doctor->id)->with('guardado','ok');
+
         }
-
-        $doctor->save();
-
-
-        return redirect()->route('secretaria.doctor.edit',$doctor->id)->with('guardado','ok');
-
 
     }
 
@@ -105,13 +113,16 @@ class DoctorSecretariaController extends Controller
      */
     public function edit($id)
     {
-        $doctor = Doctor::find($id);
-        $consultorio=Consultorio::all();
-        $vista=view('vista_paginas.secretaria.doctor.editar_info_doctor_secretaria')
-            ->with('datos_doctor',$doctor)
-        ->with('consultorios', $consultorio);
+       if(Gate::authorize('loginSecretaria')){
+           $doctor = Doctor::find($id);
+           $consultorio=Consultorio::all();
+           $vista=view('vista_paginas.secretaria.doctor.editar_info_doctor_secretaria')
+               ->with('datos_doctor',$doctor)
+               ->with('consultorios', $consultorio);
 
-        return $vista;
+           return $vista;
+
+       }
     }
 
     /**
@@ -124,41 +135,43 @@ class DoctorSecretariaController extends Controller
     public function update(Request $request, $id)
     {
 
-        $request->validate([
-            'inputImgPerfil' => ['image',],
-            'consultorio' => 'required',
-            'nombre' => 'required',
-            'apellido' => 'required',
-            'usuario' => 'required',
-            'email' => ['required', 'email',],
-            'telefono' => ['required', 'integer'],
-            'celular' => ['required', 'integer'],
-            'sexo' => 'required',
-            'status' => 'required'
-        ]);
+        if(Gate::authorize('loginSecretaria')){
 
-        $doctor=Doctor::find($id);
-        $doctor->id_consultorio=$request->input('consultorio');
-        $doctor->nombre=$request->input('nombre');
-        $doctor->apellidos=$request->input('apellido');
-        $doctor->usuario=$request->input('usuario');
-        $doctor->email=$request->input('email');
-        $doctor->telefono=$request->input('telefono');
-        $doctor->celular=$request->input('celular');
-        $doctor->sexo=$request->input('sexo');
-        $doctor->horarios=$request->input('horarios');
-        $doctor->rol="Doctor";
-        $doctor->status=$request->input('status');
+            $request->validate([
+                'inputImgPerfil' => ['image',],
+                'consultorio' => 'required',
+                'nombre' => 'required',
+                'apellido' => 'required',
+                'usuario' => 'required',
+                'email' => ['required', 'email',],
+                'telefono' => ['required', 'integer'],
+                'celular' => ['required', 'integer'],
+                'sexo' => 'required',
+                'status' => 'required'
+            ]);
 
-        if($request->hasFile('inputImgPerfil')){
-            Storage::delete($doctor->profile_photo_path);
-            $doctor->profile_photo_path=$request->file('inputImgPerfil')->store('public/fotos_perfil');
+            $doctor=Doctor::find($id);
+            $doctor->id_consultorio=$request->input('consultorio');
+            $doctor->nombre=$request->input('nombre');
+            $doctor->apellidos=$request->input('apellido');
+            $doctor->usuario=$request->input('usuario');
+            $doctor->email=$request->input('email');
+            $doctor->telefono=$request->input('telefono');
+            $doctor->celular=$request->input('celular');
+            $doctor->sexo=$request->input('sexo');
+            $doctor->horarios=$request->input('horarios');
+            $doctor->rol="Doctor";
+            $doctor->status=$request->input('status');
+
+            if($request->hasFile('inputImgPerfil')){
+                Storage::delete($doctor->profile_photo_path);
+                $doctor->profile_photo_path=$request->file('inputImgPerfil')->store('public/fotos_perfil');
+            }
+
+            $doctor->update();
+
+            return redirect()->route('secretaria.doctor.edit',$id)->with('modificado','ok');
         }
-
-        $doctor->update();
-
-        return redirect()->route('secretaria.doctor.edit',$id)->with('modificado','ok');
-
     }
 
     /**
